+++ src/modules/esm.mjs
@@ -1,0 +1,42 @@
+import fs from "fs";
+import path from "path";
+import { release, version } from "os";
+import { createServer as createServerHttp } from "http";
+import "./files/c.js";
+
+const random = Math.random();
+
+let unknownObject;
+
+const readJsonFile = (filePath) => {
+  const data = fs.readFileSync(filePath, "utf8");
+  return JSON.parse(data);
+};
+const aJsonPath = path.resolve("src/modules/files/a.json");
+const bJsonPath = path.resolve("src/modules/files/b.json");
+if (random > 0.5) {
+  unknownObject = readJsonFile(aJsonPath);
+} else {
+  unknownObject = readJsonFile(bJsonPath);
+}
+console.log(`Release: ${release()}`);
+console.log(`Version: ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${import.meta.url}`);
+console.log(`Path to current directory is ${path.dirname(import.meta.url)}`);
+
+const myServer = createServerHttp((_, res) => {
+  res.end("Request accepted");
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+  console.log(`Server is listening on port ${PORT}`);
+  console.log("To terminate it, use Ctrl+C combination");
+});
+
+export { unknownObject, myServer };
